Dev added AppTest.java

+++ GITProject08/src/main/java/com/raja/main/ArithmeticOperations.java
@@ -6,7 +6,7 @@ public class ArithmeticOperations {
 		return a+b;
 	}
 	
-	public float sub(int a, int b) {
+	public float sub(float a, float b) {
 		return a+b;
 	}
 }
